refactor: move secondary full button component to folder sign
