+++ src/pages/SignIn.tsx
@@ -5,7 +5,7 @@ import logo from "../assets/logo.svg";
 import LogoSmallScreen from "../components/Sign/LogoSmallScreen";
 import Greetings from "../components/Typography/Greetings";
 import CustomFormInput from "../components/Sign/CustomFormInput";
-import SecondaryFullButton from "../components/Button/SecondaryFullButton";
+import SecondaryFullButton from "../components/Sign/SecondaryFullButton";
 import Help from "../components/Sign/Help";
 import { complimentContent, mainContent } from "../components/Sign/Sign.style";
